added language code 'mul' from ISO-639-2;
according to wikipedia '... is applied when several languages are used
and it is not practical to specify all the appropriate language codes'

In scenarios applied on normalized treebanks, it was often not practical
to specify the language ('just process the language which you find in the document').
Now, 'mul' used in get_zone('mul') matches any language.

(However, it is not clear whether the semantics of this special code
in ISO-639-2 is 'any-language content' or rather 'mixed-language content'.
In the latter case it would be better to substitute 'mul' with some
other value ('any'?) for our purpose, but that won't be compliant with ISO.