Appositions in HamleDT (so far only for cs and grc) are treated as hypotactic structures marked with a new afun.
In PDT, apposition is treated as a paratactic structure (similarly to coordinations) governed by a comma with afun=Apos.
In HamleDT, the first member of apposition governs the (optional) comma and the second member which has afun=Apposition.
We cannot reuse afun=Apos for HamleDT because it would break several tests and assumptions.