Keep amw and tamw separate even in CoNLL-U.
Even the merged UD files will be better accepted by Github if they are smaller. But we will only split the training files which are already split (now to 4 files, from now on to 7 files) and the UD infrastructure knows how to deal with it. The dev and test portions will stay one file each.